+++ R/generateResultsTable.R
@@ -17,6 +17,14 @@
 #' obj <- enrich(
 #' compounds = input.sample, 
 #' data = FELLA.sample)
+#' 
+#' tab <- generateResultsTable(
+#' method = "hypergeom",
+#' threshold = 0.1, 
+#' object = obj, 
+#' data = FELLA.sample)
+#' head(tab)
+#' 
 #' tab <- generateResultsTable(
 #' threshold = 0.1, 
 #' object = obj, 
@@ -54,7 +62,7 @@ generateResultsTable <- function(
     if (!checkArgs$valid)
         stop("Bad argument when calling function 'generateResultsGraph'.")
     
-    if (is.na(getValid(object, method)) || !getValid(object, method)) {
+    if (is.na(getValid(object, method)) | !getValid(object, method)) {
         warning(
             paste0("Mehod ", method, " has not been executed yet. "),  
             "Returning NULL...")
@@ -70,8 +78,7 @@ generateResultsTable <- function(
         if (pscores[1] >= threshold) {
             message("No pathway is below the p-value threshold.")
         } else {
-            last <- min(plimit, tail(which(pscores < threshold), 1))
-            pscores <- pscores[1:last]
+            pscores <- head(pscores[pscores < threshold], plimit)
             
             # Info to display
             paths <- names(pscores)
@@ -92,25 +99,22 @@ generateResultsTable <- function(
             message("Done.")
             
             return(out.hypergeom)
-            
         }
-    }
-    
-    # HEAT DIFFUSION
-    if (method == "diffusion") {
-        message("Writing diffusion results...")
+    } else {
+        # Diffusion and pagerank
+        message("Writing ", method, " results...")
         
-        pscores <- sort(getPscores(object, "diffusion"))
+        pscores <- sort(getPscores(object, method))
         pscores[pscores < capPscores] <- capPscores
         
         if (pscores[1] >= threshold) {
-            message("No node is below the p-value threshold.")
+            message("No node is below the p-score threshold.")
         } else {
-            last <- min(nlimit, tail(which(pscores < threshold), 1))
-            pscores <- pscores[1:last]
+            pscores <- head(pscores[pscores < threshold], nlimit)
             pscores <- pscores[order(names(pscores))]
             
             nodeIds <- names(pscores)
+            # Take names (1st name in KEGG, abbreviate if it's too long)
             nodeNames <- as.character(sapply(nodeIds, function(id) {
                 ans <- data@keggdata@id2name[[id]]
                 if (length(ans) == 0) 
@@ -122,73 +126,19 @@ generateResultsTable <- function(
                 return(ans)
             }))
             nodeCom <- V(getGraph(data))[nodeIds]$com
-            nodeTypes <- sapply(
-                nodeCom, function(x) 
-                    switch(
-                        x, 
-                        "1" = "Pathway", 
-                        "2" = "Module", 
-                        "3" = "Enzyme", 
-                        "4" = "Reaction", 
-                        "5" = "Compound"))
-            
-            out.diffusion <- data.frame(
-                nodeIds,
-                nodeTypes,
-                nodeNames, 
-                pscores, 
-                stringsAsFactors = FALSE)[order(nodeCom), ]
-            names(out.diffusion) <- c(
-                "KEGG id", 
-                "Entry type", 
-                "KEGG name", 
-                "p.value")
-            rownames(out.diffusion) <- NULL
-            
-            message("Done.")
-            
-            return(out.diffusion)
-        }
-    }
-    
-    # PAGERANK
-    if (method == "pagerank") {
-        message("Writing pagerank results...")
-        
-        pscores <- sort(getPscores(object, "pagerank"))
-        pscores[pscores < capPscores] <- capPscores
-        
-        if (pscores[1] >= threshold) {
-            message("No node is below the p-value threshold.")
-        } else {
-            last <- min(nlimit, tail(which(pscores < threshold), 1))
-            pscores <- pscores[1:last]
-            pscores <- pscores[order(names(pscores))]
-            
-            nodeIds <- names(pscores)
-            nodeNames <- as.character(sapply(nodeIds, function(id) {
-                ans <- data@keggdata@id2name[[id]][1]
-                if (length(ans) == 0) return(NULL)
-                
-                if (nchar(ans) > LabelLengthAtPlot) 
-                    ans <- paste0(substr(ans, 1, LabelLengthAtPlot), "...")
-                return(ans)
-            }))
-            nodeCom <- V(getGraph(data))[nodeIds]$com
             nodeTypes <- listCategories()[nodeCom]
             
-            out.pagerank <- data.frame(
+            out <- data.frame(
                 "KEGG id" = nodeIds,
                 "Entry type" = nodeTypes,
                 "KEGG name" = nodeNames, 
                 "p.score" = pscores, 
                 stringsAsFactors = FALSE)[order(nodeCom), ]
-
-            rownames(out.pagerank) <- NULL
+            rownames(out) <- NULL
             
             message("Done.")
             
-            return(out.pagerank)
+            return(out)
         }
     }
 }